refactor: split up separate routes
